+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-03 16:09:05
-// Commit: feat(theme): add bi-directional event bus handler for theme
+// Updated at: 2026-08-03 22:51:25
+// Commit: chore(theme): update dependencies and cleanup unused imports in theme
 
 class themeComponentState {
-  final String commitRef = '3305';
-  final String lastMessage = 'feat(theme): add bi-directional event bus handler for theme';
-  final DateTime timestamp = DateTime.parse('2026-08-03T16:09:05+05:00');
+  final String commitRef = '3336';
+  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
+  final DateTime timestamp = DateTime.parse('2026-08-03T22:51:25+05:00');
 }